downsampling/requirements.txt
git+https://github.com/quixio/quix-streams.git@feature/sinks
prettytable
python-dotenv